requirements.txt
Flask==3.1.2
psycopg[binary]==3.2.10
PyYAML==6.0.3
requests==2.32.5
waitress==3.0.2
